refactor: move pepper/salt generation to separate common package to prevent cyclic dependencies
